requirements.txt
# Core Dependencies
torch>=2.3.0
transformers>=4.40.0
numpy>=1.24.0
pillow>=10.0.0

# Training
deepspeed>=0.14.0
peft>=0.10.0
bitsandbytes>=0.43.0
accelerate>=0.28.0
safetensors>=0.4.0

# Inference
vllm>=0.6.0
tensorrt-llm>=0.10.0

# Data Processing
datasets>=2.18.0
aiohttp>=3.9.0
aiofiles>=23.2.0
warcio>=1.7.0
trafilatura>=1.6.0
fasttext>=0.9.2
langdetect>=1.0.9
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
opencv-python>=4.9.0
scikit-learn>=1.4.0
scipy>=1.12.0

# Storage
minio>=7.2.0
psycopg2-binary>=2.9.9
redis>=5.0.0
sqlalchemy>=2.0.0

# API Server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9

# Monitoring & Logging
prometheus-client>=0.20.0
tensorboard>=2.16.0
wandb>=0.16.0

# Evaluation
pycocoevalcap>=1.0.0
nltk>=3.8.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
click>=8.1.0